refactor: make dirty cache  OS-specific
